Deleted commented code, moved skill service, deleted hydra skill service.
