--- client-side/src/app/hydra-client/services/skill/hydra-skill.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-
-import { HydraSkillService } from './hydra-skill.service';
-
-describe('HydraSkillService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [HydraSkillService]
-    });
-  });
-
-  it('should be created', inject([HydraSkillService], (service: HydraSkillService) => {
-    expect(service).toBeTruthy();
-  }));
-});
