+++ client-side/src/app/portals/Caliber/assess/assess.component.ts
@@ -21,7 +21,7 @@ import { HydraBatchService } from '../../../hydra-client/services/batch/hydra-ba
 import { HydraBatch } from '../../../hydra-client/entities/HydraBatch';
 import { HydraBatchUtilService } from '../../../services/hydra-batch-util.service';
 import { HydraTrainee } from '../../../hydra-client/entities/HydraTrainee';
-import { SkillService } from '../services/skill.service';
+import { SkillService } from '../../../hydra-client/services/skill/skill.service';
 import { Skill } from '../../../hydra-client/entities/Skill';
 
 @Component({
