+++ client-side/src/app/portals/Caliber/caliber.module.ts
@@ -29,7 +29,6 @@ import { PanelService } from './services/panel.service';
 import { GradeService } from './services/grade.service';
 import { NoteService } from './services/note.service';
 import { GranularityService } from './reports/services/granularity.service';
-import { SkillService } from './services/skill.service';
 import { TrainingTypeService } from './services/training-type.service';
 import { ColorService } from './services/colors/color.service';
 import { VpHomeLineGraphService } from './services/graph/vp-home-line-graph.service';
@@ -110,7 +109,7 @@ import { VpBarGraphComponent } from './home/vp-bar-graph/vp-bar-graph.component'
 import { VpLineGraphComponent } from './home/vp-line-graph/vp-line-graph.component';
 import { VpPanelGraphComponent } from './home/vp-panel-graph/vp-panel-graph.component';
 import { SettingsComponent } from './settings/settings.component';
-import { CategoriesComponent } from './settings/categories/categories.component';
+import { SkillsComponent } from './settings/skills/skills.component';
 import { LocationsComponent } from './settings/locations/locations.component';
 import { TrainersComponent } from './settings/trainers/trainers.component';
 import { DeactivateTrainerComponent } from './settings/trainers/deactivatetrainer/deactivatetrainer.component';
@@ -231,7 +230,7 @@ import { CategoryService } from './services/category.service';
     VpLineGraphComponent,
     VpPanelGraphComponent,
     SettingsComponent,
-    CategoriesComponent,
+    SkillsComponent,
     TrainersComponent,
     LocationsComponent,
     DeactivateTrainerComponent,
@@ -310,14 +309,12 @@ import { CategoryService } from './services/category.service';
     VpHomeSelectorService,
     ColorService,
     LocationService,
-    SkillService,
     CategoriesService,
     GranularityService,
     AlertsService,
     VpHomeBarGraphService,
     VpHomePanelGraphService,
     EvaluationService,
-    SkillService,
     TrainingTypeService,
     ReportsService,
     QCStatusService,
