+++ client-side/src/app/portals/Caliber/caliber.routes.ts
@@ -7,7 +7,7 @@ import { AssessComponent } from './assess/assess.component';
 import { ManageComponent } from './manage/manage.component';
 import { ReportsComponent } from './reports/reports.component';
 import { SettingsComponent } from './settings/settings.component';
-import { CategoriesComponent } from './settings/categories/categories.component';
+import { SkillsComponent } from './settings/skills/skills.component';
 import { LocationsComponent } from './settings/locations/locations.component';
 import { TrainersComponent } from './settings/trainers/trainers.component';
 import { DeactivateTrainerComponent } from './settings/trainers/deactivatetrainer/deactivatetrainer.component';
@@ -118,8 +118,8 @@ export const routes: Routes = [
         },
         children: [
           {
-            path: 'categories',
-            component: CategoriesComponent
+            path: 'skills',
+            component: SkillsComponent
           },
           {
             path: 'locations',
