+++ client-side/src/app/portals/Caliber/manage/manage.component.spec.ts
@@ -8,7 +8,7 @@ import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 // services
 import { BatchService } from '../services/batch.service';
 import { TrainingTypeService } from '../services/training-type.service';
-import { SkillService } from '../services/skill.service';
+import { SkillService } from '../../../hydra-client/services/skill/skill.service';
 import { LocationService } from '../services/location.service';
 import { TrainerService } from '../services/trainer.service';
 import { TraineeService } from '../services/trainee.service';
